requirements.txt
pocketsphinx-python
ai2thor
keyboard
time
sounddevice
soundfile
sys
os